backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI for GPT-4o Vision OCR
openai==2.16.0

# Qwen3-TTS dependencies
qwen-tts
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
soundfile==0.12.1
numpy>=1.24.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
aiofiles==23.2.1
